comments

+++ FreelanceWebApp/project/views.py
@@ -11,7 +11,7 @@ from project.paginations import pagination
 def project(request):
 
     template = "project/project.html"
-    object_list = Project.objects.filter(project_status__contains=1).order_by('-project_time')
+    object_list = Project.objects.filter(project_status__contains=1).order_by('-project_time')#latest projects are shown according to the latest time
 
     pages = pagination(request, object_list, 2)
 
@@ -25,17 +25,18 @@ def project(request):
     else:
         return redirect('message:welcome')
 
-def details(request, project_id):
+# viewing the project
+def details(request, project_id):#project id is passed
     template = "project/details.html"
     try:
         details = Project.objects.get(pk=project_id)
     except Project.DoesNotExist:
-        raise Http404("Project does not exist.")
+        raise Http404("Project does not exist.")# if project doesnot exist it throws http error
     context = {
         'details': details,
         'nbar': 'project',
     }
-    if request.user.is_authenticated:
+    if request.user.is_authenticated:# if the user is authenticated user is redirected to project details
         return render(request, template, context)
     else:
         return redirect('welcome')
@@ -45,7 +46,7 @@ def search(request):
     query = request.GET.get('q')
     if query:
         results = Project.objects.filter(Q(project_title__icontains=query) | Q(project_description__icontains=query))
-    else:
+    else:# if not authenticated he/she is redirected to welcome page
         results = Project.objects.filter(project_status__contains=1)
     pages = pagination(request, results, 1)
     context = {
@@ -78,6 +79,7 @@ def create(request):
     else:
         return redirect('user:login')
 
+# editing the existing project
 def update(request, project_id):
     project=Project()
     project_obj = Project.objects.get(pk=project_id)
@@ -85,32 +87,34 @@ def update(request, project_id):
         'project': project_obj,
     }
     uid = request.user.id
-    pid = project_obj.user_id
+    pid = project_obj.user_id#if user id matches with project id then update page is shown
     if uid == pid:
         return render(request, 'project/update.html', context)
-    else:
+    else:# else it stays on the current page
         return redirect('project:project')
-
+# update database
 def update_db(request, project_id):
     project_obj = Project.objects.get(pk=project_id)
     project_data = request.POST
+
+    # Gives user the chance to update
     project_obj.project_title = request.POST['title']
     project_obj.project_description = request.POST['description']
     project_obj.project_type = request.POST['pr_type']
-    project_obj.save()
-    messages.success(request, f'Post updated.')
+    project_obj.save() # changes are saved
+    messages.success(request, f'Post updated.')# message displayed by redirecting on the project page
     return redirect('project:project')
 
-
-def delete(request,pk):
+# Deleting a project
+def delete(request,pk):# Gets id of the project to be deleted
     project_obj = Project.objects.get(pk = pk)
     uid = request.user.id
     pid = project_obj.user_id
-    if uid == pid:
+    if uid == pid:# if user id and and project id matches then the project is deleted
         project_obj.delete()
         messages.success(request, f'Your post has been deleted.')
         return redirect('project:project')
-    else:
+    else:# message is displayed if the condition does not matchs
         messages.warning(request, f'Cannot delete this post')
         return redirect('project:project')
